requirements.txt
cffi==1.8.3
cryptography==1.5.2
falcon==1.0.0
falcon-jwt-checker==0.1.0
idna==2.1
py==1.4.31
pyasn1==0.1.9
pycparser==2.16
PyJWT==1.4.2
pytest==3.0.3
python-mimeparse==1.6.0
six==1.10.0
